Keep temp.cif file for fitting process

+++ easyDiffractionLib/Calculators/GSASII.py
@@ -111,7 +111,8 @@ INS  1PRCF22   0.000000E+00   0.000000E+00
         finally:
             # Clean up
             for p in pathlib.Path(os.path.dirname(self.filename)).glob("easydiffraction_temp*"):
-                p.unlink()
+                if os.path.basename(p) != "easydiffraction_temp.cif":
+                    p.unlink()
 
         if self.background is None:
             bg = np.zeros_like(this_x_array)
